Let the ctor take the same argument type as the member is (ulong)
[git-p4: depot-paths = "//depot/qt/main/": change = 49760]

+++ src/tools/quuid.h
@@ -64,7 +64,7 @@ struct Q_EXPORT QUuid
     {
 	memset( this, 0, sizeof(QUuid) );
     }
-    QUuid( uint l, ushort w1, ushort w2, uchar b1, uchar b2, uchar b3, uchar b4, uchar b5, uchar b6, uchar b7, uchar b8 )
+    QUuid( ulong l, ushort w1, ushort w2, uchar b1, uchar b2, uchar b3, uchar b4, uchar b5, uchar b6, uchar b7, uchar b8 )
     {
 	data1 = l;
 	data2 = w1;
